lecture6/mid-semester assignment/requirements.txt
apache-airflow
pandas
yfinance
feedparser
textblob
scikit-learn
joblib